Add: ✨ Alternative way to write a build design pattern
This is a much simpler version we can use to do the exactly the same
as the UserBuilder.ts class.
The difference is that the UserBuilder.ts class in principle can be much
more advanced if we have more complex fields of address and so on.
That would be the best use case, since we can use the function
name to set things to be more descriptive